# W2M_Microservice/rest/src/main/resources/OpenApiRestConfiguration.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Dekra Swagger SpaceCraft
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
paths:
  /naves:
    get:
      summary: List all spaceCraft
      operationId: listSpaceCrafts
      responses:
        '200':
          description: All SpaceCrafts
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/SpaceCraftCollectionRest"
        'default':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: create a spaceCraft
      operationId: createSpaceCraft
      requestBody:
        required: true
        content:
           application/json:
              schema:
                $ref: "#/components/schemas/SpaceCraftRest"
      responses:
        '201':
          description: Created
  /naves/{id}:
    get:
      summary: Find a specific spaceCraft
      operationId: findSpaceCraftById
      parameters:
        - name: id
          in: path
          description: unique spaceCraft identification
          required: true
          schema:
           type: integer
           format: int64
      responses:
        '200':
          description: SpaceCraft selected
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/SpaceCraftRest"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: update a exist spaceCraft
      operationId: updateSpaceCraft
      parameters:
        - name: id
          in: path
          description: unique spaceCraft identification
          required: true
          schema:
           type: integer
           format: int64
      requestBody:
        required: true
        content:
           application/json:
              schema:
                $ref: "#/components/schemas/SpaceCraftRest"
      responses:
        '201':
          description: Modified
    delete:
      summary: Delete a specific spaceCraft
      operationId: deleteSpaceCraftById
      parameters:
        - name: id
          in: path
          description: unique spaceCraft identification
          required: true
          schema:
           type: integer
           format: int64
      responses:
        '202':
          description: Accepted
  /naves/find:
    get:
      summary: List spaceCraft by filter
      operationId: listSpaceCraftsByFilter
      parameters:
        - name: name
          in: query
          description: name spaceCraft
          required: false
          schema:
           type: string
      responses:
        '200':
          description: SpaceCrafts
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/SpaceCraftCollectionRest"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    SpaceCraftRest:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string  
    SpaceCraftCollectionRest:
      type: array
      maxItems: 100
      items:
        $ref: "#/components/schemas/SpaceCraftRest"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string